+++ lib/fileManager/inc/ParserFile.hpp
@@ -24,8 +24,8 @@ namespace fileManager {
             std::vector<std::string> cleanContent(std::vector<std::string> v);
 
             /* Scearch Tool */
-            std::string findStr(std::string s, File f, unsigned int i = 0);
-            std::string findStr(std::string s, std::vector<std::string> v, unsigned int i = 0);
+            std::string findStr(std::string s, File f, size_t i = 0);
+            std::string findStr(std::string s, std::vector<std::string> v, size_t i = 0);
             std::vector<std::string> findAStr(std::string s, File f);
             std::vector<std::string> findAStr(std::string s, std::vector<std::string> v);
 
@@ -34,7 +34,7 @@ namespace fileManager {
             /* Tools clean */
             std::string cleanLine(std::string s);
             std::vector<std::string> cleanVector(std::vector<std::string> v);
-            std::string find(std::string &s, std::vector<std::string> v, unsigned int i);
+            std::string find(std::string &s, std::vector<std::string> v, size_t i);
             std::vector<std::string> findA(std::string &s, std::vector<std::string> v);
 
     };
