+++ lib/fileManager/src/ParserFile.cpp
@@ -38,7 +38,7 @@ namespace fileManager {
         return r;
     }
 
-    std::string ParserFile::find(std::string &s, std::vector<std::string> v, unsigned int i)
+    std::string ParserFile::find(std::string &s, std::vector<std::string> v, size_t i)
     {
         std::string ec;
         if (i > v.size())
@@ -56,7 +56,7 @@ namespace fileManager {
     {
         std::vector<std::string> r;
 
-        for (unsigned int i = 0; i < v.size(); i++)
+        for (size_t i = 0; i < v.size(); i++)
             if (v[i].find(s, 0) != std::string::npos)
                 r.push_back(v[i]);
         return r;
@@ -66,9 +66,9 @@ namespace fileManager {
 
     std::vector<std::string> ParserFile::cleanContent(std::vector<std::string> v) { return cleanVector(v); }
 
-    std::string ParserFile::findStr(std::string s, File f, unsigned int i) { return find(s, f.getContent(), i); }
+    std::string ParserFile::findStr(std::string s, File f, size_t i) { return find(s, f.getContent(), i); }
 
-    std::string ParserFile::findStr(std::string s, std::vector<std::string> v, unsigned int i) { return find(s, v, i); }
+    std::string ParserFile::findStr(std::string s, std::vector<std::string> v, size_t i) { return find(s, v, i); }
 
     std::vector<std::string> ParserFile::findAStr(std::string s, File f) { return findA(s, f.getContent()); }
